% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svaRetro.R
\docType{package}
\name{svaRetro}
\alias{svaRetro}
\title{svaRetro: a package for retrotransposed transcript detection}
\description{
svaRetro contains functions for detecting retrotransposed transcripts from structural variant calls.
}
\details{
For more details on the features of StructuralVariantAnnotation, read the vignette:
`browseVignettes(package = "svaRetro")`
}
